feat: add FIFTEEN_SECONDS_INTERVAL constant and refactor balance fetching logic
