+++ frontend/constants/intervals.ts
@@ -2,6 +2,8 @@ export const ONE_SECOND_INTERVAL = 1_000;
 
 export const FIVE_SECONDS_INTERVAL = 5 * ONE_SECOND_INTERVAL;
 
+export const FIFTEEN_SECONDS_INTERVAL = 15 * ONE_SECOND_INTERVAL;
+
 export const ONE_MINUTE_INTERVAL = 60 * ONE_SECOND_INTERVAL;
 
 export const FIVE_MINUTE_INTERVAL = 5 * ONE_MINUTE_INTERVAL;
